+++ src/header.rs
@@ -3,13 +3,18 @@
 use serde::{Serialize, Deserialize};
 use std::time::SystemTime;
 
+///Represents the file type
 #[derive(Serialize, Deserialize, Debug)]
 pub enum FileType {
+    ///Metadata is for a directory. (see [`std::fs::Metadata::is_dir`])
     Dir,
+    ///Metadata is for a regular file. (see [`std::fs::Metadata::is_file`])
     File,
+    ///Metadata is for a symbolic link. (see [`std::fs::Metadata::is_symlink`])
     SystemLink,
 }
 
+///A serialisable representation of file metadata
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Metadata {
     file_type: FileType,
@@ -46,5 +51,6 @@ impl From<std::fs::Metadata> for Metadata {
 }
 
 impl Metadata {
+    ///Length of the data in the file
     pub fn len(&self) -> u128 { self.size }
 }